chore: 24-09-21

+++ install.sh
@@ -5,16 +5,30 @@ DOTFILES="$HOME/dotfiles"
 
 echo "🍉 Setting up vim"
 VIMRC="$HOME/.vimrc"
-rm "$VIMRC"
+if [ -f "$VIMRC" ]; then
+  rm "$VIMRC"
+fi
 ln -s "$DOTFILES"/.vimrc "$VIMRC"
 
 if [ "$(uname)" = "Darwin" ]; then
+  echo "🍉 Setting up mac"
+  # shellcheck source=/dev/null
   . "$DOTFILES"/macos/install.sh
 fi
 
+echo "🍉 Setting up zsh"
+NPC="$HOME/.non_public_commands.sh"
+echo "🍉     Setting up non publich commands"
+if [ -f "$NPC" ]; then
+  rm "$NPC"
+fi
+ln -s "$DOTFILES"/zsh/.non_public_commands.sh "$NPC"
+# shellcheck source=/dev/null
 . "$DOTFILES"/zsh/install.sh
 
 echo "🍉 Setting up git"
 GITCONFIG="$HOME/.gitconfig"
-rm "$GITCONFIG"
+if [ -f "$GITCONFIG" ]; then
+  rm "$GITCONFIG"
+fi
 ln -s "$DOTFILES"/.gitconfig "$GITCONFIG"
